Extract logic into ValidationAnnotationUtils
